# docs/openapi.yaml
openapi: 3.1.0
swagger: "2.0"
info:
  title: Vehicle Tracking API
  version: 1.0.0
  description: API for managing vehicle tracking, maintenance schedules, and shared access
  contact:
    name: API Support
    email: support@vehicletracking.com
  license:
    name: MIT
servers:
  - url: https://api.vehicletracking.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Local development server
paths:
...
  /vehicles:
    post:
      tags: [Vehicles]
      summary: Add new vehicle
      description: Register a new vehicle for tracking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        201:
          description: Vehicle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        400:
          description: Invalid vehicle data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Vehicles]
      summary: List vehicles
      description: Get list of vehicles with maintenance status
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VehicleResponse'
        401:
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /vehicles/{id}:
    put:
      tags: [Vehicles]
      summary: Update vehicle
      description: Update vehicle details and maintenance intervals
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        200:
          description: Vehicle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        400:
          description: Invalid vehicle data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Vehicles]
      summary: Delete vehicle
      description: Permanently delete a vehicle and its associated data
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Vehicle deleted successfully
        401:
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        404:
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained during authentication
  schemas:
    DamageStatus:
      type: string
      enum: [NONE, MINOR, MODERATE, SEVERE]
      example: "NONE"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid token"
        code:
          type: string
          example: "INVALID_TOKEN"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "expiresInHours"
              message:
                type: string
                example: "must be between 1 and 168"
...
    VehicleRequest:
      type: object
      required: [plate, type]
      properties:
        plate:
          type: string
          pattern: "^([A-Z]{2}[\\s-]?[0-9]{3}[\\s-]?[A-Z]{2}|[A-Z]{2}[\\s-]?[0-9]{4}[A-Z]{2})$"
          example: "AB123CD"
        type:
          type: string
          enum: ["car", "motorcycle", "truck", "military"]
          example: "car"
        insuranceInterval:
          type: integer
          minimum: 1
          example: 365
        taxInterval:
          type: integer
          minimum: 1
          example: 365
        inspectionInterval:
          type: integer
          minimum: 1
          example: 365
        damageStatus:
          $ref: '#/components/schemas/DamageStatus'
        damageDescription:
          type: string
          nullable: true
          example: "Scratch on rear bumper"
    VehicleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "d7d98f80-05f3-4110-9af9-3f5c7e0b8ac1"
        plate:
          type: string
          example: "AB123CD"
        type:
          type: string
          example: "car"
        insuranceDue:
          type: string
          format: date-time
          nullable: true
          example: "2025-02-27T23:38:35Z"
        taxDue:
          type: string
          format: date-time
          nullable: true
          example: "2025-02-27T23:38:35Z"
        inspectionDue:
          type: string
          format: date-time
          nullable: true
          example: "2025-02-27T23:38:35Z"
        damageStatus:
          $ref: '#/components/schemas/DamageStatus'
        damageDescription:
          type: string
          nullable: true
          example: "Scratch on rear bumper"
        status:
          type: object
          properties:
            insurance:
              type: string
              enum: ["red", "orange", "green", "gray"]
              example: "green"
            tax:
              type: string
              enum: ["red", "orange", "green", "gray"]
              example: "green"
            inspection:
              type: string
              enum: ["red", "orange", "green", "gray"]
              example: "green"
      example:
        id: "d7d98f80-05f3-4110-9af9-3f5c7e0b8ac1"
        plate: "AB123CD"
        type: "car"
        insuranceDue: "2025-02-27T23:38:35Z"
        taxDue: "2025-02-27T23:38:35Z"
        inspectionDue: "2025-02-27T23:38:35Z"
        damageStatus: "NONE"
        damageDescription: null
        status:
          insurance: "green"
          tax: "green"
          inspection: "green"
...